+++ Calculate.java
@@ -2,10 +2,10 @@ import java.util.Arrays;
 
 public class Calculate {
 	public static void main(String[] args) {
+		CalculateMethod method = CalculateMethod.of(args[0]);
 		String[] methodArgs = Arrays.copyOfRange(args, 1, args.length);
 
-		CalculateAction.of(args[0])
-			.map(method -> method.calculate(methodArgs))
-			.ifPresent(System.out::println);
+		String result = method.calculate(methodArgs);
+		System.out.println(result);
 	}
 }
