+++ README.md
@@ -1,39 +1,15 @@
 # Sample CLI Calculator
 Sample repository to demonstrate Git and GitHub
 
-## Requirements
-- JDK 8
-
 ## Usage:
 ```bash
-$ java Calculate <action-name> <arguments...>
+$ java Calculate <method> <arguments...>
 ```
 
-### Available Actions:
+### Available Methods:
 - `add A B` - Adds `A` and `B`
 - `subtract A B` - Subtracts the `B` from `A`
 - `multiply A B` - Multiplies the `A` by `B`
 - `divide A B` - Divides `A` by `B`
 - `mod A B` - Performs modulo operations from integer `A` and `B`
-- `deg A` - Converts radians to degrees from integer `A`
-- `rad A` - Converts degrees to radians from integer `A`
-- `abs A` - Converts an integer `A` to its absolute value
-- `deg2gon A` - Converts degrees `A` to a gon
-- `rad2gon A` - Converts radians `A` to a gon
-- `sin rad` - Gets the sin of the inputted value in radians
-- `cos rad` - Gets the cos of the inputted value in radians
-- `tan rad` - Gets the tan of the inputted value in radians
-- `csc rad` - Gets the csc of the inputted value in radians
-- `sec rad` - Gets the sec of the inputted value in radians
-- `cot rad` - Gets the cot of the inputted value in radians
-- `pow B E` - Performs exponential operation with base `B` and exponent `E`
-- `sqrt A` - Performs square root operation on `A`
-- `cbrt A` - Performs cube root operation on `A`
-- `root A R` - Gets the `R`th root of value `A`
-- `hypotenuse A B` - Computes the hypotenuse from `A` and `B`
-- `sqarea A` - Computes the area of a square from a side `A`
-- `triarea A B` - Computes the area of a triangle from  `A` and `B` (base and height)
-- `base2 A` - Converts `A` Decimal value to its Binary value
-- `base8 A` - Converts `A` Decimal value to its Octal value
-- `base16 A` - Converts `A` Decimal value to its Hexadecimal value
-- `round A B` - Round off `A` to the nearest `B` value
+
