--- build.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-rm *.class
-javac *.java
-
